feat: moving Router setup to AppRoutes component
